@@ -1,6 +1,6 @@
 import 'dart:math' as math;
 
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 import 'package:zflutter/zflutter.dart';
 
 const illoSize = 128;
@@ -30,7 +30,7 @@ class OnTheGo extends StatefulWidget {
 }
 
 class _OnTheGoState extends State<OnTheGo> with TickerProviderStateMixin {
-  AnimationController animationController;
+  late AnimationController animationController;
 
   @override
   void initState() {
@@ -138,7 +138,7 @@ class Dude extends StatelessWidget {
   }
 
   Widget rightThigh(Color thighColor, Color lineColor,
-      {ZVector legRotation, bool lineOposite = false}) {
+      {ZVector legRotation = ZVector.zero, bool lineOposite = false}) {
     Widget shoe(bool isRight) {
       final shoelace = ZShape(
         path: [
@@ -173,6 +173,7 @@ class Dude extends StatelessWidget {
                 ZPositioned(
                   translate: ZVector.only(y: 4),
                   child: ZShape(
+                    color: Colors.transparent,
                     visible: false,
                   ),
                 )
@@ -605,6 +606,7 @@ class Lady extends StatelessWidget {
 
     // z-sort hack
     final zSortHack = new ZShape(
+      color: Colors.transparent,
       path: [ZMove.vector(ZVector.only(z: 16))],
       visible: false,
     );
