@@ -1,15 +1,14 @@
+//@dart=2.12
 import 'package:flutter/material.dart';
 import 'package:zflutter/src/core/core.dart';
 
 class ZRenderer {
-  Path path = Path();
-
   Paint paint = Paint()
     ..isAntiAlias = true
     ..strokeCap = StrokeCap.round
     ..strokeJoin = StrokeJoin.round;
 
-  final Canvas canvas;
+  final Canvas? canvas;
 
   ZRenderer(this.canvas);
 
@@ -17,26 +16,45 @@ class ZRenderer {
     paint.strokeCap = value;
   }
 
+  void stroke(Path path, Color color, double lineWidth) {
+    paint.color = color;
+    paint.strokeWidth = lineWidth;
+    paint.style = PaintingStyle.stroke;
+    canvas?.drawPath(path, paint);
+  }
+
+  void fill(Path path, Color color) {
+    paint.color = color;
+    paint.style = PaintingStyle.fill;
+    canvas?.drawPath(path, paint);
+  }
+}
+
+class ZPathBuilder {
+  Path path = Path();
+
+  ZPathBuilder();
+
   void begin() {
     path.reset();
   }
 
   void move(ZVector point) {
-    path.moveTo(point.x!, point.y!);
+    path.moveTo(point.x, point.y);
   }
 
   void line(ZVector point) {
-    path.lineTo(point.x!, point.y!);
+    path.lineTo(point.x, point.y);
   }
 
   void bezier(ZVector cp0, ZVector cp1, ZVector end) {
     path.cubicTo(
-      cp0.x!,
-      cp0.y!,
-      cp1.x!,
-      cp1.y!,
-      end.x!,
-      end.y!,
+      cp0.x,
+      cp0.y,
+      cp1.x,
+      cp1.y,
+      end.x,
+      end.y,
     );
   }
 
@@ -55,33 +73,17 @@ class ZRenderer {
     path.close();
   }
 
-  void renderToPath(Path path, List<ZPathCommand> pathCommands, bool isClosed) {
-    var previousPath = this.path;
-    this.path = path;
-    renderPath(pathCommands, isClosed: isClosed);
-    this.path = previousPath;
-  }
-
-  void renderPath(List<ZPathCommand> pathCommands, {bool isClosed = false}) {
+  Path renderPath(List<ZPathCommand> pathCommands, {bool isClosed = false}) {
     begin();
+    ZVector previousPoint = ZVector.zero;
     pathCommands.forEach((it) {
-      it.render(this);
+      it.render(this, previousPoint);
+      previousPoint = it.endRenderPoint;
     });
+
     if (isClosed) {
       closePath();
     }
-  }
-
-  void stroke(Color color, double lineWidth) {
-    paint.color = color;
-    paint.strokeWidth = lineWidth;
-    paint.style = PaintingStyle.stroke;
-    canvas.drawPath(path, paint);
-  }
-
-  void fill(Color color) {
-    paint.color = color;
-    paint.style = PaintingStyle.fill;
-    canvas.drawPath(path, paint);
+    return path;
   }
 }
