@@ -24,19 +24,21 @@ class GettingStartedSamples {
     title: 'Animated',
     route: '/animated',
     builder: (_) => Spin(
-        builder: (context, rotate) => ZIllustration(
-              zoom: 2,
-              children: [
-                ZPositioned(
-                  rotate: rotate,
-                  child: ZCircle(
-                    diameter: 80,
-                    stroke: 20,
-                    color: Color(0xFFCC2255),
-                  ),
-                ),
-              ],
-            )),
+      builder: (context, rotate) => ZIllustration(
+        zoom: 2,
+        children: [
+
+          ZPositioned(
+            rotate: rotate,
+            child: ZCircle(
+              diameter: 80,
+              stroke: 20,
+              color: Color(0xFFCC2255),
+            ),
+          ),
+        ],
+      ),
+    ),
   );
 
   static final Example drag = Example(
@@ -81,8 +83,6 @@ class GettingStartedSamples {
       },
     ),
   );
-
-
 
   static List<Example> get list => [zoom, animated, drag, boxAdapter];
 }
