@@ -1,6 +1,7 @@
+//@dart=2.12
+//import 'dart:ui';
+import 'package:vector_math/vector_math_64.dart' as vector;
 import 'dart:ui';
-
-import 'package:quiver/core.dart';
 
 import 'dart:math' as math;
 
@@ -11,6 +12,7 @@ export 'render/render_box.dart';
 export 'render/render_shape.dart';
 export 'path_command.dart';
 export 'renderer.dart';
+export 'path_builder.dart';
 
 export 'widgets/box_adapter.dart';
 
@@ -19,19 +21,16 @@ const tau = math.pi * 2;
 // A immutable 3D vector
 // Todo: Add unit test for this. It is important
 class ZVector {
-  final double? x;
-  final double? y;
-  final double? z;
+  final double x;
+  final double y;
+  final double z;
 
-  const ZVector(this.x, this.y, this.z)
-      : assert(x != null && y != null);
+  const ZVector(this.x, this.y, this.z);
 
-  const ZVector.only({this.x = 0, this.y = 0, this.z = 0})
-      : assert(x != null && y != null);
+  const ZVector.only({this.x = 0, this.y = 0, this.z = 0});
 
-  const ZVector.all(double? value)
-      : assert(value != null),
-        this.x = value,
+  const ZVector.all(double value)
+      : this.x = value,
         this.y = value,
         this.z = value;
   static const ZVector zero = ZVector.all(0);
@@ -39,22 +38,22 @@ class ZVector {
 
   bool operator ==(v) => v is ZVector && x == v.x && y == v.y && z == v.z;
 
-  int get hashCode => hash3(x.hashCode, y.hashCode, z.hashCode);
+  int get hashCode => Object.hashAll([x, y, z]);
 
   ZVector add({double x = 0, double y = 0, double z = 0}) {
-    return ZVector(this.x! + x, this.y! + y, this.z! + z);
+    return ZVector(this.x + x, this.y + y, this.z + z);
   }
 
   ZVector subtract({double x = 0, double y = 0, double z = 0}) {
-    return ZVector(this.x! - x, this.y! - y, this.z! - z);
+    return ZVector(this.x - x, this.y - y, this.z - z);
   }
 
   ZVector subtractVector(ZVector v) {
-    return ZVector(this.x! - v.x!, this.y! - v.y!, this.z! - v.z!);
+    return ZVector(this.x - v.x, this.y - v.y, this.z - v.z);
   }
 
   ZVector addVector(ZVector v) {
-    return ZVector(this.x! + v.x!, this.y! + v.y!, this.z! + v.z!);
+    return ZVector(this.x + v.x, this.y + v.y, this.z + v.z);
   }
 
   ZVector rotate(ZVector? rotation) {
@@ -63,20 +62,20 @@ class ZVector {
     return this.rotateZ(rotation.z).rotateY(rotation.y).rotateX(rotation.x);
   }
 
-  ZVector rotateZ(double? angle) {
-    return _rotateProperty(angle, Axis.x, Axis.y);
+  ZVector rotateZ(double angle) {
+    return _rotateProperty(angle, VectorAxis.x, VectorAxis.y);
   }
 
-  ZVector rotateX(double? angle) {
-    return _rotateProperty(angle, Axis.y, Axis.z);
+  ZVector rotateX(double angle) {
+    return _rotateProperty(angle, VectorAxis.y, VectorAxis.z);
   }
 
-  ZVector rotateY(double? angle) {
-    return _rotateProperty(angle, Axis.x, Axis.z);
+  ZVector rotateY(double angle) {
+    return _rotateProperty(angle, VectorAxis.x, VectorAxis.z);
   }
 
-  ZVector _rotateProperty(double? angle, Axis propA, Axis propB) {
-    if (angle == null || angle % tau == 0) {
+  ZVector _rotateProperty(double angle, VectorAxis propA, VectorAxis propB) {
+    if (angle % tau == 0) {
       return this;
     }
     var cos = math.cos(angle);
@@ -88,57 +87,52 @@ class ZVector {
         {propA: a * cos - b * sin, propB: b * cos + a * sin});
   }
 
-  replaceAxisInMap(Map<Axis, double> axis) {
-    double? x, y, z;
-    axis.forEach((key, value) {
-      if (key == Axis.x) x = value;
-      if (key == Axis.y) y = value;
-      if (key == Axis.z) z = value;
-    });
+  ZVector replaceAxisInMap(Map<VectorAxis, double> axis) {
+    double? x = axis[VectorAxis.x];
+    double? y = axis[VectorAxis.y];
+    double? z = axis[VectorAxis.z];
+
     return ZVector(x ?? this.x, y ?? this.y, z ?? this.z);
   }
 
-  Map<Axis, double?> get toMap => {
-        Axis.x: x,
-        Axis.y: y,
-        Axis.z: z,
-      };
+  Map<VectorAxis, double> get toMap =>
+      {VectorAxis.x: x, VectorAxis.y: y, VectorAxis.z: z};
 
   ZVector multiply(ZVector? scale) {
     if (scale == null) return this;
-    final mx = scale.x ?? 1;
-    final my = scale.y ?? 1;
-    final mz = scale.z ?? 1;
-    return ZVector(x! * mx, y! * my, z! * mz);
+    final mx = scale.x;
+    final my = scale.y;
+    final mz = scale.z;
+    return ZVector(x * mx, y * my, z * mz);
   }
 
   ZVector divide(ZVector? scale) {
     if (scale == null) return this;
-    final mx = scale.x ?? 1;
-    final my = scale.y ?? 1;
-    final mz = scale.z ?? 1;
-    return ZVector(x! / mx, y! / my, z! / mz);
+    final mx = scale.x;
+    final my = scale.y;
+    final mz = scale.z;
+    return ZVector(x / mx, y / my, z / mz);
   }
 
   ZVector multiplyScalar(num? scale) {
     if (scale == null) return this;
     final m = scale;
-    return ZVector(x! * m, y! * m, z! * m);
+    return ZVector(x * m, y * m, z * m);
   }
 
   ZVector transform(ZVector translation, ZVector rotation, ZVector scale) {
     return this.multiply(scale).rotate(rotation).addVector(translation);
   }
 
-  static ZVector lerp(ZVector a, ZVector? b, double t) {
-    final x = lerpDouble(a.x, b?.x ?? 0.0, t);
-    final y = lerpDouble(a.y, b?.y ?? 0.0, t);
-    final z = lerpDouble(a.z, b?.z ?? 0.0, t);
-    return ZVector(x, y, z);
+  static ZVector lerp(ZVector? a, ZVector? b, double t) {
+    final x = lerpDouble(a?.x, b?.x ?? 0.0, t);
+    final y = lerpDouble(a?.y, b?.y ?? 0.0, t);
+    final z = lerpDouble(a?.z, b?.z ?? 0.0, t);
+    return ZVector(x!, y!, z!);
   }
 
   double magnitude() {
-    var sum = this.x! * this.x! + this.y! * this.y! + this.z! * this.z!;
+    var sum = this.x * this.x + this.y * this.y + this.z * this.z;
     return getMagnitudeSqrt(sum);
   }
 
@@ -151,7 +145,7 @@ class ZVector {
   }
 
   double magnitude2d() {
-    var sum = this.x! * this.x! + this.y! * this.y!;
+    var sum = this.x * this.x + this.y * this.y;
     return getMagnitudeSqrt(sum);
   }
 
@@ -173,12 +167,12 @@ class ZVector {
 
   /// Cross product.
   ZVector cross(ZVector other) {
-    final double _x = x!;
-    final double _y = y!;
-    final double _z = z!;
-    final double ox = other.x!;
-    final double oy = other.y!;
-    final double oz = other.z!;
+    final double _x = x;
+    final double _y = y;
+    final double _z = z;
+    final double ox = other.x;
+    final double oy = other.y;
+    final double oz = other.z;
     return ZVector.only(
       x: _y * oz - _z * oy,
       y: _z * ox - _x * oz,
@@ -188,13 +182,44 @@ class ZVector {
 
   ZVector unit() {
     var total = magnitude();
-    return ZVector(x! / total, y! / total, z! / total);
+    return ZVector(x / total, y / total, z / total);
   }
 
   @override
   String toString() {
     return 'V($x, $y, $z)';
   }
+
+  vector.Vector3 asVector3() {
+    return vector.Vector3(x, y, z);
+  }
+
+  ZVector applyMatrix4(vector.Matrix4 arg) {
+    final argStorage = arg.storage;
+    final v0 = x;
+    final v1 = y;
+    final v2 = z;
+    return ZVector(
+      argStorage[0] * v0 +
+          argStorage[4] * v1 +
+          argStorage[8] * v2 +
+          argStorage[12],
+      argStorage[1] * v0 +
+          argStorage[5] * v1 +
+          argStorage[9] * v2 +
+          argStorage[13],
+      argStorage[2] * v0 +
+          argStorage[6] * v1 +
+          argStorage[10] * v2 +
+          argStorage[14],
+    );
+  }
 }
 
-enum Axis { x, y, z }
+extension ZVector3 on vector.Vector3 {
+  ZVector asVector() {
+    return ZVector(x, y, z);
+  }
+}
+
+enum VectorAxis { x, y, z }
