@@ -1,110 +1,90 @@
-import 'core.dart';
+//@dart=2.12
 
+import 'core.dart';
+import 'package:vector_math/vector_math_64.dart' as vector;
 // TODO: This Paths needs to be immutable;
 
 abstract class ZPathCommand {
-  final ZVector? endRenderPoint = ZVector.zero;
+  final ZVector endRenderPoint = ZVector.zero;
 
-  void reset();
+  const ZPathCommand();
 
-  ZPathCommand transform(ZVector translation, ZVector rotate, ZVector scale);
+  ZPathCommand transform(vector.Matrix4 transformation);
 
-  void render(ZRenderer renderer);
+  void render(
+    ZPathBuilder renderer,
+    ZVector previousPoint,
+  );
 
-  ZVector? point({index = 0});
+  ZVector point({index = 0});
 
-  ZVector? renderPoint({int index = 0});
-
-  set previous(ZVector? previousPoint) {}
-
-  ZPathCommand clone();
+  ZVector renderPoint({int index = 0});
 }
 
 class ZMove extends ZPathCommand {
-  ZVector? _point;
+  final ZVector _point;
 
-  ZVector? _renderPoint;
+  ZVector get endRenderPoint => _point;
 
-  ZVector? get endRenderPoint => _renderPoint;
+  const ZMove.vector(this._point);
 
-  ZMove.vector(ZVector this._point) {
-    _renderPoint = _point!.copy();
+  ZMove(double x, double y, double z) : _point = ZVector(x, y, z);
+
+  ZMove.only({double x = 0, double y = 0, double z = 0})
+      : _point = ZVector(x, y, z);
+
+  ZPathCommand transform(vector.Matrix4 transformation) {
+    return ZMove.vector(_point.applyMatrix4(transformation));
   }
 
-  ZMove(double x, double y, double z) {
-    _renderPoint = ZVector(x, y, z);
+  void render(ZPathBuilder renderer, ZVector previousPoint) {
+    renderer.move(_point);
   }
 
-  ZMove.only({double x = 0, double y = 0, double z = 0}) {
-    _renderPoint = ZVector(x, y, z);
-  }
-
-  void reset() {
-    _renderPoint = _point;
-  }
-
-  ZPathCommand transform(ZVector translation, ZVector rotate, ZVector scale) {
-    return ZMove.vector(_renderPoint!.transform(translation, rotate, scale));
-  }
-
-  void render(ZRenderer renderer) {
-    renderer.move(_renderPoint!);
-  }
-
-  ZVector? point({index = 0}) {
+  ZVector point({index = 0}) {
     return _point;
   }
 
-  ZVector? renderPoint({index = 0}) {
-    return _renderPoint;
-  }
-
-  ZPathCommand clone() {
-    return ZMove.vector(this.point()!);
+  ZVector renderPoint({index = 0}) {
+    return _point;
   }
 }
 
 class ZLine extends ZPathCommand {
-  ZVector? _point;
+  late ZVector _point;
 
-  ZVector? _renderPoint;
+  late ZVector _renderPoint;
 
-  ZVector? get endRenderPoint => _renderPoint;
+  ZVector get endRenderPoint => _renderPoint;
 
-  ZLine.vector(ZVector this._point) {
-    _renderPoint = _point!.copy();
+  ZLine.vector(this._point) {
+    _renderPoint = _point.copy();
   }
 
   ZLine(double x, double y, double z) {
+    _point = ZVector(x, y, z);
     _renderPoint = ZVector(x, y, z);
   }
 
   ZLine.only({double x = 0, double y = 0, double z = 0}) {
+    _point = ZVector(x, y, z);
     _renderPoint = ZVector(x, y, z);
   }
 
-  void reset() {
-    _renderPoint = _point;
+  ZPathCommand transform(vector.Matrix4 transformation) {
+    return ZLine.vector(_renderPoint.applyMatrix4(transformation));
   }
 
-  ZPathCommand transform(ZVector translation, ZVector rotate, ZVector scale) {
-    return ZLine.vector(_renderPoint!.transform(translation, rotate, scale));
+  void render(ZPathBuilder renderer, ZVector previousPoint) {
+    renderer.line(_renderPoint);
   }
 
-  void render(ZRenderer renderer) {
-    renderer.line(_renderPoint!);
-  }
-
-  ZVector? point({index = 0}) {
+  ZVector point({index = 0}) {
     return _point;
   }
 
-  ZVector? renderPoint({index = 0}) {
+  ZVector renderPoint({index = 0}) {
     return _renderPoint;
-  }
-
-  ZPathCommand clone() {
-    return ZLine.vector(_point!);
   }
 }
 
@@ -119,19 +99,13 @@ class ZBezier extends ZPathCommand {
     renderPoints = points.map((e) => e.copy()).toList();
   }
 
-  void reset() {
-    /* renderPoints.asMap().forEach((index, vector) {
-      vector.set(points[index]);
-    });*/
-  }
-
-  ZPathCommand transform(ZVector translation, ZVector rotate, ZVector scale) {
+  ZPathCommand transform(vector.Matrix4 transformation) {
     return ZBezier(renderPoints.map((point) {
-      return point.transform(translation, rotate, scale);
+      return point.applyMatrix4(transformation);
     }).toList());
   }
 
-  void render(ZRenderer renderer) {
+  void render(ZPathBuilder renderer, ZVector previousPoint) {
     renderer.bezier(renderPoints[0], renderPoints[1], renderPoints[2]);
   }
 
@@ -142,29 +116,22 @@ class ZBezier extends ZPathCommand {
   ZVector renderPoint({index = 0}) {
     return renderPoints[index];
   }
-
-  ZPathCommand clone() {
-    return ZBezier(this.points);
-  }
 }
 
 const double _arcHandleLength = 9 / 16;
 
 class ZArc extends ZPathCommand {
   late List<ZVector> points;
-  ZVector? _previous = ZVector.zero;
 
   late List<ZVector> renderPoints;
 
   ZVector get endRenderPoint => renderPoints.last;
 
-  ZArc.list(this.points, [this._previous]) {
+  ZArc.list(this.points, [ZVector? previous]) {
     renderPoints = points.map((e) => e.copy()).toList();
   }
 
-  ZArc({required ZVector corner, required ZVector end, ZVector? previous}) {
-    _previous = previous;
-
+  ZArc({required ZVector corner, required ZVector end}) {
     points = [corner, end];
 
     renderPoints = points.map((e) => e.copy()).toList();
@@ -176,15 +143,14 @@ class ZArc extends ZPathCommand {
     renderPoints = List.generate(renderPoints.length, (i) => points[i]);
   }
 
-  ZPathCommand transform(ZVector translation, ZVector rotate, ZVector scale) {
+  ZPathCommand transform(vector.Matrix4 transformation) {
     return ZArc.list(renderPoints.map((point) {
-      return point.transform(translation, rotate, scale);
+      return point.applyMatrix4(transformation);
     }).toList());
   }
 
-  void render(ZRenderer renderer) {
-    assert(_previous != null);
-    var prev = _previous!;
+  void render(ZPathBuilder renderer, ZVector previousPoint) {
+    var prev = previousPoint;
     var corner = renderPoints[0];
     var end = renderPoints[1];
     var a = ZVector.lerp(prev, corner, _arcHandleLength);
@@ -198,15 +164,5 @@ class ZArc extends ZPathCommand {
 
   ZVector renderPoint({index = 0}) {
     return renderPoints[index];
-  }
-
-  @override
-  set previous(ZVector? previousPoint) {
-    assert(previousPoint != null);
-    _previous = previousPoint;
-  }
-
-  ZPathCommand clone() {
-    return ZArc.list(points, _previous);
   }
 }
