@@ -1,30 +1,113 @@
+//@dart=2.12
+
+import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 
 import '../core.dart';
 
-class RenderZBox extends RenderBox {
-  double? sortValue = 0;
+abstract class RenderZBox extends RenderBox {
+  bool _debugSortedValue = false;
+  bool _debugTransformedValue = false;
 
-  ZVector origin = ZVector.zero;
+  double sortValue = 0;
 
-  void performSort() {
-    sortValue = this.origin.z;
+  @override
+  @mustCallSuper
+  void performLayout() {
+    _debugTransformedValue = false;
+    _buildMatrix();
+    performTransformation();
+    _debugTransformedValue = true;
+    sort();
+  }
+
+  Matrix4? _matrix;
+  Matrix4 get matrix {
+    assert(_matrix != null, 'Matrix accessed before performing layout');
+    return _matrix!;
+  }
+
+  _buildMatrix() {
+    final anchorParentData = parentData;
+
+    _matrix = Matrix4.identity();
+    if (anchorParentData is ZParentData) {
+      anchorParentData.transforms.forEach((transform) {
+        final matrix4 = Matrix4.translationValues(transform.translate.x,
+            transform.translate.y, transform.translate.z);
+
+        matrix4.rotateX(transform.rotate.x);
+        matrix4.rotateY(-transform.rotate.y);
+        matrix4.rotateZ(transform.rotate.z);
+
+        matrix4.scale(transform.scale.x, transform.scale.y, transform.scale.z);
+        matrix..multiply(matrix4);
+      });
+    }
   }
 
   @override
-  Size computeDryLayout(BoxConstraints constraints) => constraints.biggest;
+  bool get sizedByParent => true;
+
+  void performTransformation();
+
+  void performSort();
+
+  int compareSort(RenderZBox renderBox) {
+    return sortValue.compareTo(renderBox.sortValue);
+  }
+
+  @mustCallSuper
+  void sort() {
+    _debugSortedValue = true;
+    performSort();
+  }
+
+  @override
+  void paint(PaintingContext context, Offset offset) {
+    assert(_debugSortedValue, 'requires sorted value');
+    debugTransformed();
+    super.paint(context, offset);
+  }
+
+  void debugTransformed() {
+    assert(_debugTransformedValue, 'requires transformation to be performed');
+  }
+
+  @override
+  Size computeDryLayout(BoxConstraints constraints) {
+    return constraints.biggest;
+  }
+
+  @override
+  void performResize() {
+    size = constraints.biggest;
+    assert(size.isFinite);
+  }
 }
 
-enum SortMode { inherit, stack, update }
+enum SortMode {
+  // Each child inside the group is sorted by its own center
+  // The group acts as a proxy
+  inherit,
+  // Children are sorted following the order in the list
+  stack,
+  // Children are encapsulated and painted in the order described
+  // The group is painted by
+  update,
+}
 
-class RenderZMultiChildBox extends RenderZBox
+class RenderMultiChildZBox extends RenderZBox
     with
         ContainerRenderObjectMixin<RenderZBox, ZParentData>,
         RenderBoxContainerDefaultsMixin<RenderZBox, ZParentData> {
-  RenderZMultiChildBox({
+  RenderMultiChildZBox({
     List<RenderZBox>? children,
-    SortMode sortMode = SortMode.inherit,
-  })  : this.sortMode = sortMode {
+    SortMode? sortMode = SortMode.inherit,
+    ZVector? sortPoint,
+  })  : assert(sortMode != null),
+        this.sortMode = sortMode,
+        _sortPoint = sortPoint {
     addAll(children);
   }
 
@@ -40,129 +123,124 @@ class RenderZMultiChildBox extends RenderZBox
   }
 
   @override
-  void performLayout() {
-    super.performLayout();
+  bool get sizedByParent => true;
 
+  @override
+  void performTransformation() {
     final BoxConstraints constraints = this.constraints;
 
     RenderZBox? child = firstChild;
 
     while (child != null) {
-      final ZParentData? childParentData = child.parentData as ZParentData?;
-
-      if (child is RenderZMultiChildBox && child.sortMode == SortMode.inherit) {
+      final ZParentData childParentData = child.parentData as ZParentData;
+      if (child is RenderMultiChildZBox && child.sortMode == SortMode.inherit) {
         child.layout(constraints, parentUsesSize: true);
       } else {
         child.layout(constraints, parentUsesSize: true);
       }
-
-      /*  final Size childSize = child.size;
-        width = math.max(width, childSize.width);
-        height = math.max(height, childSize.height);*/
-
-      child = childParentData?.nextSibling;
+      child = childParentData.nextSibling;
     }
-    performSort();
   }
 
-  @override
-  void performResize() {
-    size = constraints.biggest;
-    assert(size.isFinite);
+  ZVector? get sortPoint => _sortPoint;
+  ZVector? _sortPoint;
+  set sortPoint(ZVector? value) {
+    if (value == sortPoint) return;
+    _sortPoint = value;
+    markNeedsLayout();
   }
+
+  List<RenderZBox>? sortedChildren;
 
   @override
   void performSort() {
+    final children = _getFlatChildren();
     if (sortMode == SortMode.stack || sortMode == SortMode.update) {
-      final children = getFlatChildren();
-      sortValue = children.fold(0,
-              (dynamic previousValue, element) => previousValue + element.sortValue) /
-          children.length;
-    } else {
-      super.performSort();
+      if (sortPoint != null) {
+        sortValue = _sortPoint!.applyMatrix4(matrix).z;
+      } else {
+        sortValue = children.fold<double>(0, (previousValue, element) {
+              return (previousValue + element.sortValue);
+            }) /
+            children.length;
+      }
     }
+    if (sortMode == SortMode.update) {
+      children..sort((a, b) => a.compareSort(b));
+    }
+    sortedChildren = children;
   }
 
-  @override
-  bool get sizedByParent => true;
+  SortMode? sortMode;
 
-  SortMode sortMode;
-
-  List<RenderZBox> getFlatChildren() {
+  List<RenderZBox> _getFlatChildren() {
     List<RenderZBox> children = [];
 
     RenderZBox? child = firstChild;
 
     while (child != null) {
-      final ZParentData? childParentData = child.parentData as ZParentData?;
+      final ZParentData childParentData = child.parentData as ZParentData;
 
-      if (child is RenderZMultiChildBox && child.sortMode == SortMode.inherit) {
-        children.addAll(child.getFlatChildren());
+      if (child is RenderMultiChildZBox && child.sortMode == SortMode.inherit) {
+        children.addAll(child._getFlatChildren());
       } else {
         children.add(child);
       }
-
-      /*  final Size childSize = child.size;
-        width = math.max(width, childSize.width);
-        height = math.max(height, childSize.height);*/
-
-      child = childParentData?.nextSibling;
+      child = childParentData.nextSibling;
     }
     return children;
   }
 
-  ZVector rotation = ZVector.zero;
-  ZVector scale = ZVector.zero;
-  ZVector translate = ZVector.zero;
-
   @override
   void paint(PaintingContext context, Offset offset) {
+    assert(sortMode != null);
     if (sortMode == SortMode.inherit) return;
-    List<RenderZBox> children = getFlatChildren();
-    //List<RenderBox> children = getChildrenAsList()
-    if (sortMode != SortMode.stack) {
-      children..sort((a, b) => a.sortValue!.compareTo(b.sortValue!));
+
+    for (final child in sortedChildren!) {
+      context.paintChild(child, offset);
     }
-    for (final child in children) {
-      final ZParentData childParentData = child.parentData as ZParentData;
-      context.paintChild(child, childParentData.offset + offset);
+  }
+
+  bool defaultHitTestChildren(BoxHitTestResult result,
+      {required Offset position}) {
+    if (sortMode == SortMode.inherit) return false;
+    // The x, y parameters have the top left of the node's box as the origin.
+    List<RenderZBox> children = sortedChildren!;
+
+    for (final child in children.reversed) {
+      final bool isHit = child.hitTest(result, position: position);
+
+      if (isHit) return true;
     }
+    return false;
+  }
+
+  @override
+  bool hitTest(BoxHitTestResult result, {required Offset position}) {
+    if (hitTestChildren(result, position: position) || hitTestSelf(position)) {
+      result.add(BoxHitTestEntry(this, position));
+      return true;
+    }
+    return false;
+  }
+
+  @override
+  bool hitTestChildren(BoxHitTestResult result, {required Offset position}) {
+    return defaultHitTestChildren(result, position: position);
   }
 }
 
 /// Parent data for use with [ZRenderer].
 class ZParentData extends ContainerBoxParentData<RenderZBox> {
-  ZVector translate = ZVector.zero;
-  ZVector rotate = ZVector.zero;
-  ZVector scale = ZVector.identity;
-
   List<ZTransform> transforms;
 
   ZParentData({
-    this.rotate = ZVector.zero,
-    this.scale = ZVector.identity,
-    this.translate = ZVector.zero,
     List<ZTransform>? transforms,
   }) : this.transforms = transforms ?? [];
 
-  ZParentData clone() => ZParentData(
-      rotate: rotate,
-      scale: scale,
-      translate: translate,
-      transforms: List<ZTransform>.from(transforms));
-
-  /* String toString() {
-    final List<String> values = <String>[
-      if (top != null) 'top=${debugFormatDouble(top)}',
-      if (right != null) 'right=${debugFormatDouble(right)}',
-      if (bottom != null) 'bottom=${debugFormatDouble(bottom)}',
-      if (left != null) 'left=${debugFormatDouble(left)}',
-      if (width != null) 'width=${debugFormatDouble(width)}',
-      if (height != null) 'height=${debugFormatDouble(height)}',
-    ];
-    if (values.isEmpty)
-      values.add('not positioned');
-    values.add(super.toString());
-    return values.join('; ');
-  }*/
+  ZParentData clone() {
+    return ZParentData(
+      transforms: List<ZTransform>.from(transforms),
+    );
+  }
 }
