@@ -1,7 +1,7 @@
+//@dart=2.12
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-
-import '../core.dart';
+import 'package:zflutter/zflutter.dart';
 
 /// A zWidget that paints a shape in a 3D space.
 ///
@@ -25,16 +25,12 @@ import '../core.dart';
 ///
 /// If no path is provided a dot will be painted with the stroke as diameter.
 ///
-class ZShape extends SingleChildRenderObjectWidget with ZWidget {
-  /// The path that will define the shape of the Widget
-  /// It is an ordered list of path commands : [ZMove], [ZLine], [ZArc] & [ZBezier]
-  /// See some prebuilt shapes as examples:  [ZRect], [ZRounderRect], [ZEllipse]
-  final List<ZPathCommand>? path;
-
+abstract class ZShapeBuilder extends SingleChildRenderObjectWidget
+    with ZWidget {
   /// The color of the shape. If [stroke] is more than 0, the path will be painted
   /// with a stroke of this color. If [fill] is true, it will paint the inside of
   /// the path with this color
-  final Color? color;
+  final Color color;
 
   /// The width that will be used to paint the path
   final double stroke;
@@ -57,43 +53,53 @@ class ZShape extends SingleChildRenderObjectWidget with ZWidget {
   /// If false the shape won't be painted
   final bool visible;
 
-  ZShape({
+  /// Position used for z-sorting
+  /// If null the center of the path will be used instead
+  final ZVector? sortPoint;
+
+  const ZShapeBuilder({
     Key? key,
-    this.path,
-    this.color,
+    required this.color,
     this.front = const ZVector.only(z: 1),
     this.backfaceColor,
     this.stroke = 1,
     this.closed = true,
     this.fill = false,
     this.visible = true,
+    this.sortPoint,
   })  : assert(stroke >= 0),
         super(key: key);
+
+  /// The path that will define the shape of the Widget
+  /// It is an ordered list of path commands : [ZMove], [ZLine], [ZArc] & [ZBezier]
+  /// See some prebuilt shapes as examples:  [ZRect], [ZRounderRect], [ZEllipse]
+  PathBuilder buildPath();
 
   @override
   RenderZShape createRenderObject(BuildContext context) {
     return RenderZShape(
-      color: color,
-      path: path ?? [],
-      stroke: stroke,
-      close: closed,
-      fill: fill,
-      visible: visible,
-      backfaceColor: backfaceColor,
-      front: front,
-    );
+        color: color,
+        pathBuilder: buildPath(),
+        stroke: stroke,
+        close: closed,
+        fill: fill,
+        visible: visible,
+        backfaceColor: backfaceColor,
+        front: front,
+        sortPoint: sortPoint);
   }
 
   @override
   void updateRenderObject(BuildContext context, RenderZShape renderObject) {
     renderObject..color = color;
-    renderObject..path = path ?? [];
+    renderObject..pathBuilder = buildPath();
     renderObject..stroke = stroke;
     renderObject..close = closed;
     renderObject..fill = fill;
     renderObject..backfaceColor = backfaceColor;
     renderObject..front = front;
     renderObject..visible = visible;
+    renderObject..sortPoint = sortPoint;
   }
 
   @override
@@ -110,4 +116,37 @@ class ZShape extends SingleChildRenderObjectWidget with ZWidget {
   @override
   ZSingleChildRenderObjectElement createElement() =>
       ZSingleChildRenderObjectElement(this);
+}
+
+class ZShape extends ZShapeBuilder {
+  /// The path that will define the shape of the Widget
+  /// It is an ordered list of path commands : [ZMove], [ZLine], [ZArc] & [ZBezier]
+  /// See some prebuilt shapes as examples:  [ZRect], [ZRounderRect], [ZEllipse]
+  final PathBuilder path;
+
+  ZShape({
+    Key? key,
+    List<ZPathCommand>? path,
+    required Color color,
+    Color? backfaceColor,
+    double stroke = 1,
+    bool fill = false,
+    ZVector front = const ZVector.only(z: 1),
+    bool visible = true,
+    bool closed = true,
+  })  : path = SimplePathBuilder(path ?? const []),
+        assert(stroke >= 0),
+        super(
+          key: key,
+          color: color,
+          backfaceColor: backfaceColor,
+          stroke: stroke,
+          closed: closed,
+          fill: fill,
+          front: front,
+          visible: visible,
+        );
+
+  @override
+  PathBuilder buildPath() => path;
 }
