@@ -1,6 +1,5 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-
 import 'package:zflutter/zflutter.dart';
 
 class ZCylinder extends StatelessWidget {
@@ -10,8 +9,8 @@ class ZCylinder extends StatelessWidget {
   final double stroke;
   final bool fill;
 
-  final Color? color;
-  final bool? visible;
+  final Color color;
+  final bool visible;
 
   final Color? backface;
   final Color? frontface;
@@ -21,8 +20,8 @@ class ZCylinder extends StatelessWidget {
     this.length = 1,
     this.stroke = 1,
     this.fill = true,
-    this.color,
-    this.visible,
+    required this.color,
+    this.visible = true,
     this.backface,
     this.frontface,
   });
@@ -75,53 +74,72 @@ class ZCylinder extends StatelessWidget {
   }
 }
 
-class _ZCylinderMiddle extends ZShape {
-  final double? diameter;
+class _ZCylinderMiddle extends ZShapeBuilder {
+  final double diameter;
+
+  final List<ZPathCommand> path;
 
   _ZCylinderMiddle(
-      {this.diameter, List<ZPathCommand>? path, double stroke = 1, Color? color})
-      : super(path: path, stroke: stroke, color: color);
+      {required this.diameter,
+      required this.path,
+      double stroke = 1,
+      required Color color})
+      : super(stroke: stroke, color: color);
 
   @override
   RenderZCylinder createRenderObject(BuildContext context) {
     return RenderZCylinder(
-        path: path!, stroke: stroke, diameter: diameter, color: color);
+      pathBuilder: buildPath(),
+      stroke: stroke,
+      diameter: diameter,
+      color: color,
+    );
   }
 
   @override
   void updateRenderObject(BuildContext context, RenderZCylinder renderObject) {
     renderObject.diameter = diameter;
     renderObject.stroke = stroke;
-    renderObject.path = path;
+    renderObject.pathBuilder = buildPath();
     renderObject.color = color;
+  }
+
+  @override
+  PathBuilder buildPath() {
+    return SimplePathBuilder(path);
   }
 }
 
 class RenderZCylinder extends RenderZShape {
-  double? _diameter;
+  double _diameter;
 
-  double? get diameter => _diameter;
+  double get diameter => _diameter;
 
-  set diameter(double? value) {
+  set diameter(double value) {
     if (_diameter == value) return;
     _diameter = value;
 
-    markNeedsPaint();
+    markNeedsLayout();
   }
 
-  RenderZCylinder(
-      {required List<ZPathCommand> path, double? diameter, double? stroke, Color? color})
-      : _diameter = diameter,
-        super(path: path, stroke: stroke, color: color);
+  RenderZCylinder({
+    required PathBuilder pathBuilder,
+    required double diameter,
+    required double stroke,
+    required Color color,
+  })  : _diameter = diameter,
+        super(pathBuilder: pathBuilder, stroke: stroke, color: color);
 
   @override
+  bool get needsDirection => true;
+  @override
   void render(ZRenderer renderer) {
+    final builder = ZPathBuilder()..renderPath(transformedPath);
     var scale = normalVector.magnitude();
-    var strokeWidth = diameter! * scale + stroke!;
+    var strokeWidth = diameter * scale + stroke;
 
     renderer.setLineCap(StrokeCap.butt);
-    renderer.renderPath(transformedPath!);
-    renderer.stroke(color!, strokeWidth);
+    renderer.stroke(builder.path, color, strokeWidth);
     renderer.setLineCap(StrokeCap.round);
     super.render(renderer);
   }
