@@ -2,30 +2,46 @@ import 'package:flutter/widgets.dart';
 import 'package:zflutter/zflutter.dart';
 import 'dart:math' as math;
 
-class ZRect extends ZShape {
-  final double width;
-  final double height;
-
-  ZRect({
+class ZRect extends ZShapeBuilder {
+  const ZRect({
     Key? key,
     required this.width,
     required this.height,
-    Color? color,
+    required Color color,
     Color? backfaceColor,
     double stroke = 1,
     bool fill = false,
     ZVector front = const ZVector.only(z: 1),
-  })  : super(
-            key: key,
-            color: color,
-            backfaceColor: backfaceColor,
-            stroke: stroke,
-            closed: true,
-            fill: fill,
-            front: front,
-            path: performPath(width, height));
+  }) : super(
+          key: key,
+          color: color,
+          backfaceColor: backfaceColor,
+          stroke: stroke,
+          closed: true,
+          fill: fill,
+          front: front,
+          sortPoint: ZVector.zero,
+        );
 
-  static List<ZPathCommand> performPath(double width, double height) {
+  final double width;
+
+  final double height;
+
+  @override
+  PathBuilder buildPath() {
+    return RectPathBuilder(width, height);
+  }
+}
+
+class RectPathBuilder extends PathBuilder {
+  final double width;
+
+  final double height;
+
+  RectPathBuilder(this.width, this.height);
+
+  @override
+  List<ZPathCommand> buildPath() {
     final x = width / 2;
     final y = height / 2;
     return [
@@ -35,9 +51,16 @@ class ZRect extends ZShape {
       ZLine.vector(ZVector.only(x: -x, y: y))
     ];
   }
+
+  @override
+  bool shouldRebuildPath(covariant PathBuilder oldPathBuilder) {
+    return !(oldPathBuilder is RectPathBuilder) ||
+        oldPathBuilder.height != height ||
+        oldPathBuilder.width != width;
+  }
 }
 
-class ZRoundedRect extends ZShape {
+class ZRoundedRect extends ZShapeBuilder {
   final double width;
   final double height;
   final double borderRadius;
@@ -47,23 +70,39 @@ class ZRoundedRect extends ZShape {
     required this.width,
     required this.height,
     required this.borderRadius,
-    Color? color,
+    required Color color,
     Color? backfaceColor,
     double stroke = 1,
     bool fill = false,
     ZVector front = const ZVector.only(z: 1),
-  })  : super(
-            key: key,
-            color: color,
-            backfaceColor: backfaceColor,
-            stroke: stroke,
-            closed: true,
-            fill: fill,
-            front: front,
-            path: performPath(width, height, borderRadius));
+  }) : super(
+          key: key,
+          color: color,
+          backfaceColor: backfaceColor,
+          stroke: stroke,
+          closed: true,
+          fill: fill,
+          front: front,
+          sortPoint: ZVector.zero,
+        );
 
-  static List<ZPathCommand> performPath(
-      double width, double height, double borderRadius) {
+  @override
+  PathBuilder buildPath() {
+    return RoundedRectPathBuilder(width, height, borderRadius);
+  }
+}
+
+class RoundedRectPathBuilder extends PathBuilder {
+  final double width;
+
+  final double height;
+
+  final double borderRadius;
+
+  RoundedRectPathBuilder(this.width, this.height, this.borderRadius);
+
+  @override
+  List<ZPathCommand> buildPath() {
     var xA = width / 2;
     var yA = height / 2;
     var shortSide = math.min(xA, yA);
@@ -121,9 +160,17 @@ class ZRoundedRect extends ZShape {
 
     return path;
   }
+
+  @override
+  bool shouldRebuildPath(covariant PathBuilder oldPathBuilder) {
+    return !(oldPathBuilder is RoundedRectPathBuilder) ||
+        oldPathBuilder.height != height ||
+        oldPathBuilder.width != width ||
+        oldPathBuilder.borderRadius != borderRadius;
+  }
 }
 
-class ZCircle extends ZShape {
+class ZCircle extends ZShapeBuilder {
   final double diameter;
 
   final int quarters;
@@ -132,7 +179,7 @@ class ZCircle extends ZShape {
     Key? key,
     required this.diameter,
     this.quarters = 4,
-    Color? color,
+    required Color color,
     bool closed = false,
     Color? backfaceColor,
     double stroke = 1,
@@ -140,17 +187,27 @@ class ZCircle extends ZShape {
     ZVector front = const ZVector.only(z: 1),
   })  : assert(quarters >= 0 && quarters <= 4),
         super(
-            key: key,
-            color: color,
-            backfaceColor: backfaceColor,
-            stroke: stroke,
-            closed: closed,
-            fill: fill,
-            front: front,
-            path: ZEllipse.performPath(diameter, diameter, quarters));
+          key: key,
+          color: color,
+          backfaceColor: backfaceColor,
+          stroke: stroke,
+          closed: closed,
+          fill: fill,
+          front: front,
+          sortPoint: ZVector.zero,
+        );
+
+  @override
+  PathBuilder buildPath() {
+    return EllipsePathBuilder(
+      height: diameter,
+      width: diameter,
+      quarters: quarters,
+    );
+  }
 }
 
-class ZEllipse extends ZShape {
+class ZEllipse extends ZShapeBuilder {
   final double width;
   final double height;
 
@@ -161,71 +218,96 @@ class ZEllipse extends ZShape {
     required this.width,
     required this.height,
     this.quarters = 4,
-    Color? color,
+    required Color color,
     Color? backfaceColor,
     double stroke = 1,
     bool fill = false,
     ZVector front = const ZVector.only(z: 1),
   })  : assert(quarters >= 0 && quarters <= 4),
         super(
-            key: key,
-            color: color,
-            backfaceColor: backfaceColor,
-            stroke: stroke,
-            closed: false,
-            fill: fill,
-            front: front,
-            path: performPath(width, height, quarters));
+          key: key,
+          color: color,
+          backfaceColor: backfaceColor,
+          stroke: stroke,
+          closed: false,
+          fill: fill,
+          front: front,
+          sortPoint: ZVector.zero,
+        );
 
-  static List<ZPathCommand> performPath(
-      double width, double height, int quarters) {
+  @override
+  PathBuilder buildPath() {
+    return EllipsePathBuilder(
+      height: height,
+      width: width,
+      quarters: quarters,
+    );
+  }
+}
+
+class EllipsePathBuilder extends PathBuilder {
+  final double width;
+  final double height;
+
+  final int quarters;
+
+  EllipsePathBuilder({
+    required this.width,
+    required this.height,
+    required this.quarters,
+  });
+
+  @override
+  List<ZPathCommand> buildPath() {
     var x = width / 2;
     var y = height / 2;
 
     var path = [
       ZLine.vector(ZVector.only(x: 0, y: -y)),
-      ZArc.list(
-        [
-          ZVector.only(x: x, y: -y),
-          ZVector.only(x: x, y: 0),
-        ],
-        null,
+      ZArc(
+        corner: ZVector.only(x: x, y: -y),
+        end: ZVector.only(x: x, y: 0),
       ),
     ];
 
     if (quarters > 1) {
-      path.add(ZArc.list(
-        [
-          ZVector.only(x: x, y: y),
-          ZVector.only(x: 0, y: y),
-        ],
-        null,
-      ));
+      path.add(
+        ZArc(
+          corner: ZVector.only(x: x, y: y),
+          end: ZVector.only(x: 0, y: y),
+        ),
+      );
     }
     if (quarters > 2) {
-      path.add(ZArc.list(
-        [
-          ZVector.only(x: -x, y: y),
-          ZVector.only(x: -x, y: 0),
-        ],
-        null,
-      ));
+      path.add(
+        ZArc(
+          corner: ZVector.only(x: -x, y: y),
+          end: ZVector.only(x: -x, y: 0),
+        ),
+      );
     }
     if (quarters > 3) {
-      path.add(ZArc.list(
-        [
-          ZVector.only(x: -x, y: -y),
-          ZVector.only(x: 0, y: -y),
-        ],
-        null,
-      ));
+      path.add(
+        ZArc(
+          corner: ZVector.only(x: -x, y: -y),
+          end: ZVector.only(x: 0, y: -y),
+        ),
+      );
     }
 
     return path;
   }
+
+  @override
+  bool shouldRebuildPath(covariant PathBuilder oldPathBuilder) {
+    return !(oldPathBuilder is EllipsePathBuilder) ||
+        oldPathBuilder.height != height ||
+        oldPathBuilder.width != width ||
+        oldPathBuilder.quarters != quarters;
+  }
 }
 
-class ZPolygon extends ZShape {
+class ZPolygon extends ZShapeBuilder {
   final int sides;
   final double radius;
 
@@ -233,7 +315,7 @@ class ZPolygon extends ZShape {
     Key? key,
     required this.sides,
     required this.radius,
-    Color? color,
+    required Color color,
     Color? backfaceColor,
     double stroke = 1,
     bool fill = false,
@@ -241,21 +323,45 @@ class ZPolygon extends ZShape {
   })  : assert(sides > 2),
         assert(radius > 0),
         super(
-            key: key,
-            color: color,
-            backfaceColor: backfaceColor,
-            stroke: stroke,
-            closed: true,
-            fill: fill,
-            front: front,
-            path: performPath(sides, radius));
+          key: key,
+          color: color,
+          backfaceColor: backfaceColor,
+          stroke: stroke,
+          closed: true,
+          fill: fill,
+          front: front,
+          sortPoint: ZVector.zero,
+        );
 
-  static List<ZPathCommand> performPath(int sides, double radius) {
+  @override
+  PathBuilder buildPath() {
+    return PolygonPathBuilder(sides: sides, radius: radius);
+  }
+}
+
+class PolygonPathBuilder extends PathBuilder {
+  final int sides;
+  final double radius;
+
+  PolygonPathBuilder({
+    required this.sides,
+    required this.radius,
+  });
+
+  @override
+  List<ZPathCommand> buildPath() {
     return List.generate(sides, (index) {
       final double theta = index / sides * tau - tau / 4;
       final double x = math.cos(theta) * radius;
       final double y = math.sin(theta) * radius;
       return ZLine.vector(ZVector.only(x: x, y: y));
     });
+  }
+
+  @override
+  bool shouldRebuildPath(covariant PathBuilder oldPathBuilder) {
+    return !(oldPathBuilder is PolygonPathBuilder) ||
+        oldPathBuilder.sides != sides ||
+        oldPathBuilder.radius != radius;
   }
 }
