@@ -6,7 +6,14 @@ import 'package:zflutter/src/core/widgets/widget.dart';
 class ZIllustration extends ZMultiChildWidget {
   final double zoom;
 
+  /// Whether clipBehavioring children should be clipped. See [clipBehavior].
+  ///
+  /// Some children in a stack might clipBehavior its box. When this flag is set to
+  /// [clipBehavior.clip], children cannot paint outside of the stack's box.
+  final Clip clipBehavior;
+
   ZIllustration({
+    this.clipBehavior = Clip.hardEdge,
     required List<Widget> children,
     this.zoom = 1,
   })  : assert(zoom >= 0),
@@ -24,27 +31,58 @@ class ZIllustration extends ZMultiChildWidget {
   }
 }
 
-class RenderZIllustration extends RenderZMultiChildBox {
-  double? _zoom = 1;
+class RenderZIllustration extends RenderMultiChildZBox {
+  RenderZIllustration({
+    Clip clipBehavior = Clip.none,
+    double zoom = 0,
+    List<RenderZBox>? children,
+  })  : assert(zoom >= 0),
+        _zoom = zoom,
+        _clipBehavior = clipBehavior,
+        super(children: children, sortMode: SortMode.update);
 
-  double? get zoom => _zoom;
-
-  set zoom(double? value) {
-    assert(_zoom != null && _zoom! >= 0);
+  double _zoom = 1;
+  double get zoom => _zoom;
+  set zoom(double value) {
+    assert(_zoom >= 0);
     if (_zoom == value) return;
     _zoom = value;
     markNeedsPaint();
   }
 
-  RenderZIllustration({
-    double? zoom,
-    List<RenderZBox>? children,
-  })  : assert(zoom != null && zoom >= 0),
-        _zoom = zoom,
-        super(children: children, sortMode: SortMode.update);
+  @override
+  bool get sizedByParent => true;
+
+  /// Whether clipBehavioring children should be clipped. See [clipBehavior].
+  ///
+  /// Some children in a stack might clipBehavior its box. When this flag is set to
+  /// [clipBehavior.clip], children cannot paint outside of the stack's box.
+  Clip get clipBehavior => _clipBehavior;
+  Clip _clipBehavior;
+  set clipBehavior(Clip value) {
+    if (_clipBehavior != value) {
+      _clipBehavior = value;
+      markNeedsPaint();
+    }
+  }
 
   @override
   void paint(PaintingContext context, Offset offset) {
+    if (_clipBehavior != Clip.none) {
+      context.pushClipRect(
+        needsCompositing,
+        offset,
+        Offset.zero & size,
+        paintIllustration,
+        clipBehavior: _clipBehavior,
+      );
+    } else {
+      paintIllustration(context, offset);
+    }
+  }
+
+  @protected
+  void paintIllustration(PaintingContext context, Offset offset) {
     Matrix4 matrix = Matrix4.identity()
       ..translate(size.width / 2, size.height / 2)
       ..scale(zoom, zoom, zoom);
@@ -54,45 +92,33 @@ class RenderZIllustration extends RenderZMultiChildBox {
   }
 
   @override
+  void applyPaintTransform(RenderObject child, Matrix4 transform) {
+    Matrix4 matrix = Matrix4.identity()
+      ..translate(size.width / 2, size.height / 2)
+      ..scale(zoom, zoom, zoom);
+    transform.multiply(matrix);
+  }
+
+  @override
   bool get isRepaintBoundary => true;
 
-  // TODO: Work on hitTest
   @override
-  bool hitTest(BoxHitTestResult result, {required Offset position}) {
-    assert(() {
-      if (!hasSize) {
-        if (debugNeedsLayout) {
-          throw FlutterError.fromParts(<DiagnosticsNode>[
-            ErrorSummary(
-                'Cannot hit test a render box that has never been laid out.'),
-            describeForError(
-                'The hitTest() method was called on this RenderBox'),
-            ErrorDescription(
-                "Unfortunately, this object's geometry is not known at this time, "
-                'probably because it has never been laid out. '
-                'This means it cannot be accurately hit-tested.'),
-            ErrorHint('If you are trying '
-                'to perform a hit test during the layout phase itself, make sure '
-                "you only hit test nodes that have completed layout (e.g. the node's "
-                'children, after their layout() method has been called).'),
-          ]);
-        }
-        throw FlutterError.fromParts(<DiagnosticsNode>[
-          ErrorSummary('Cannot hit test a render box with no size.'),
-          describeForError('The hitTest() method was called on this RenderBox'),
-          ErrorDescription(
-              'Although this node is not marked as needing layout, '
-              'its size is not set.'),
-          ErrorHint('A RenderBox object must have an '
-              'explicit size before it can be hit-tested. Make sure '
-              'that the RenderBox in question sets its size during layout.'),
-        ]);
-      }
-      return true;
-    }());
-    if (size.contains(position)) {
-      return true;
-    }
-    return false;
+  bool hitTestSelf(Offset position) {
+    final Rect bounds = position & size;
+    return bounds.contains(position);
+  }
+
+  @override
+  bool hitTestChildren(BoxHitTestResult result, {required Offset position}) {
+    Matrix4 matrix = Matrix4.identity()
+      ..translate(size.width / 2, size.height / 2)
+      ..scale(zoom, zoom, zoom);
+    return result.addWithPaintTransform(
+      transform: matrix,
+      position: position,
+      hitTest: (BoxHitTestResult result, Offset position) {
+        return super.hitTestChildren(result, position: position);
+      },
+    );
   }
 }
